+++ src/repo/watcher.rs
@@ -1,3 +1,5 @@
+use crate::model::{playlist::PlaylistType, watcher::Watcher};
+
 pub struct WatcherRepo {
     ctx: crate::context::AppContext,
 }
@@ -7,49 +9,85 @@ impl WatcherRepo {
         Self { ctx }
     }
 
-    /// Try to find the ID of the playlist a user has configured to watch.
-    pub fn get_watched_playlist_id_by_user_id(
-        &self,
-        user_id: &str,
-    ) -> crate::Result<Option<String>> {
-        Ok(self
-            .ctx
-            .db
-            .get()?
-            .prepare("SELECT playlist_id FROM watchers WHERE user_id = ? LIMIT 1")?
-            .query_row(&[user_id], |row| Ok(row.get(0)?))
-            .ok())
-    }
-
     /// Get all configured watchers.
-    pub fn get_all_watchers(&self) -> crate::Result<Vec<(String, String, String)>> {
+    pub fn get_all_watchers(&self) -> crate::Result<Vec<Watcher>> {
         self.ctx
             .db
-            .get() ?
-            .prepare("SELECT watchers.user_id, watchers.playlist_id, users.token FROM watchers LEFT JOIN users ON watchers.user_id = users.user_id")?
-            .query_map([], |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?)))?
+            .get()?
+            .prepare(
+                "SELECT users.id, users.token, watchers.from_playlist, watchers.to_playlist, watchers.should_remove FROM watchers
+                LEFT JOIN users ON users.id = watchers.user_id",
+            )?
+            .query_map([], |row| {
+                Ok(Watcher::try_from_row_data(
+                    row.get(0)?,
+                    row.get(1)?,
+                    row.get(2)?,
+                    row.get(3)?,
+                    row.get::<usize, u8>(4)? != 0,
+                ).unwrap())
+            })?
+            .collect::<rusqlite::Result<Vec<_>>>()
+            .map_err(|err| err.into())
+    }
+
+    /// Get all watchers for a given user ID.
+    pub fn get_all_watchers_by_user(&self, user_id: &str) -> crate::Result<Vec<Watcher>> {
+        self.ctx
+            .db
+            .get()?
+            .prepare(
+                "SELECT users.token, watchers.from_playlist, watchers.to_playlist, watchers.should_remove FROM watchers
+                LEFT JOIN users ON users.id = watchers.user_id
+                AND watchers.user_id = ?",
+            )?
+            .query_map(&[user_id], |row| {
+                Ok(Watcher::try_from_row_data(
+                    user_id.to_owned(),
+                    row.get(0)?,
+                    row.get(1)?,
+                    row.get(2)?,
+                    row.get::<usize, u8>(3)? != 0,
+                ).unwrap())
+            })?
             .collect::<rusqlite::Result<Vec<_>>>()
             .map_err(|err| err.into())
     }
 
     /// Create a watcher for a user and playlist.
-    pub fn create_watcher(&self, user_id: &str, playlist_id: &str) -> crate::Result<()> {
+    pub fn create_watcher(
+        &self,
+        user_id: &str,
+        from: PlaylistType,
+        to: PlaylistType,
+    ) -> crate::Result<()> {
         self.ctx
             .db
             .get()?
-            .prepare("INSERT INTO watchers (user_id, playlist_id) VALUES (?, ?)")?
-            .execute(&[user_id, playlist_id])?;
+            .prepare(
+                "INSERT INTO watchers
+                (user_id, from_playlist, to_playlist, should_remove, created_at)
+                VALUES (?, ?, ?, 1, datetime())",
+            )?
+            .execute(&[user_id, &from.to_string(), &to.to_string()])?;
 
         Ok(())
     }
 
-    /// Delete a user's watcher.
-    pub fn delete_watcher(&self, user_id: &str) -> crate::Result<()> {
+    /// Delete a watcher given user and playlist IDs.
+    pub fn delete_watcher(
+        &self,
+        user_id: &str,
+        from: PlaylistType,
+        to: PlaylistType,
+    ) -> crate::Result<()> {
         self.ctx
             .db
             .get()?
-            .prepare("DELETE FROM watchers WHERE user_id = ?")?
-            .execute(&[user_id])?;
+            .prepare(
+                "DELETE FROM watchers WHERE user_id = ? AND from_playlist = ? to_playlist = ?",
+            )?
+            .execute(&[user_id, &from.to_string(), &to.to_string()])?;
 
         Ok(())
     }